Support lora training and inference.
